G7-B2: one authoritative log prototype -- 2,420 declarations, 6 real defects fixed
AUTHORITATIVE SIGNATURES, and how they were established. There are THREE app entry
points, not four: DEBUG_PRINT is Ghidra's name for the SAME ADDRESS as log_message
(0x0007dda4), already bound by an __asm__ alias.
  debug_print  0x00019c70              void debug_print(uintptr_t format, ...)
  log_message/DEBUG_PRINT 0x0007dda4   void log_message(uintptr_t format, ...)
  printk (app/net) 0x0007e2fa/0x01039722  void printk(uintptr_t format, ...)
Variadic with exactly one fixed parameter: all three open push {r0,r1,r2,r3} then
ldr rX,[rY],#4 -- a post-increment load taking one word from the save area and starting
the va_list at the next. No VFP register in any prologue. void: ALL 2,526 bl/b.w sites
targeting the three addresses in app_update.bin were disassembled and not one reads r0
before writing it (the 10 that mention it do eors r0,r0). Cross-checked against
REVIEWED_CALL_ARITIES_BY_FORMAT, which keys arity by (callee, format) -- only meaningful
for a variadic callee. uintptr_t is ABI-identical to the shipped const char * here;
const char * would have emitted ~4,000 diagnostics through the SDK's __printf_like(1,2)
and changed zero instructions. printk in both links is stock libzephyr.a(printk.c.obj),
so g1_log.h guards its own declaration on #ifndef ZEPHYR_INCLUDE_SYS_PRINTK_H_ and lets
the SDK win where its header is in scope.

CONSOLIDATED: 2,420 declarations across 1,683 files in 6 trees replaced by one include
of the new recon/headers/g1_log.h, plus the 4 definition files pointed at it.

THE PAYOFF -- and it is NOT what was predicted. ZERO call sites were passing floats
under the wrong convention (the one place floats reach a sink, fuel_gauge_update
@0x10c14 via __aeabi_f2d into r2:r3 and the stack, already had a variadic declaration).
What the wrong prototypes were actually doing was INVENTING RETURN VALUES AND ARGUMENTS:
6 files / 15 call sites, found because the authoritative declaration refused to compile
them -- get_lux_info, set_imu_thread_delay, send_data_in_ble_chunks,
exit_dashboard_burial_point, test_mode_apply_base_status_cmd, reflash_fb_data_to_lcd_ex.
Each fix is evidenced by the format string's specifier count plus the shipped register
loads, and each emitted body was read back against the shipped disassembly (several
instruction-for-instruction identical).

BYTE-IDENTITY VS RE-PROOF: objects app 3,407/3,430 byte-identical, net 1,168/1,170. Of
the 23 app differences: 6 defect fixes, 14 declaration-driven (all semantics-neutral --
literal-pool dedup and scheduling), 1 generated isr_tables, 2 stock SDK units that also
differ on the net side with no related change. cfg_verify on the canonical bodies 5/5
PASS. THE 6TH, FUN_00031dd8, FAILS CORRECTED AND PASSES DEFECTIVE -- the harness compares
4 arguments at an argument-free format and the 4th is 0x20019EF3, dead scratch. Fixing
that needs a REVIEWED_CALL_ARITIES_BY_FORMAT entry, i.e. a tools/ change forbidden in
this batch, so the canonical body KEEPS ITS DEFECT and the build tree carries the fix,
documented in 3.1.

GATES ALL PASS: gen_retained_sources --check clean; both cores exit 0; nm -u 0/0;
duplicate globals 0/0; no --allow-multiple-definition/--defsym/weak/numeric-root hacks;
check_ram_pin_collisions 0/0 both cores; check_net_raw_literals 0/0;
check_thread_create_stack_args 10/10; verify_net_stock_data_window PROVEN (0 words
differing for any non-flash-pointer reason).
FLASH app 948,812 B = 96.57%, delta -16 B; net 225,581 B, delta 0. RAM delta 0 both
cores. Warnings app 1,798 -> 1,887 (+89, all -Wint-conversion on argument 1 at sites
passing a pointer expression; no instruction changes), net unchanged.
FRAMEBUFFERS -- all four byte-identical, and HELD ACROSS A -16 B IMAGE CHANGE so it is a
measurement rather than a tautology: dashboard p2_render 19b1f24a... 2,923 px
(78,211)-(564,338); dashboard p1_boot 0c5cc90b... 0 px; navigation p2_render b26c73b3...
1,098 px; navigation p1_boot 1d617c65... 656 px; spim_a p1_boot stream sha f91505ab...

DEFERRED WITH REASONS: (1) 140 zero-argument call sites in 45 build-tree files (1,071
across all trees) -- debug_print() with no arguments cannot compile against a
one-fixed-parameter prototype, and this is newly quantified as a REAL DEFECT CLASS: the
image loads r0/r1/r2 before every one of those bl's (e.g. handle_attitude_trigger @0xfec0
hoists the loads above the cbz). Needs its own per-site batch. (2) The five canonical
parity trees (1,435 declarations) -- tools/parity/recon.py builds a TEMPORARY TU whose
CFLAGS carry no recon/headers, and its -std=gnu11 deliberately preserves unprototyped
f(), which the zero-arg sites rely on; unblocks with one include-path line in tools/.
(3) FUN_00031dd8 canonical body; 8 macro-redirect files; 2 mixed-declarator statements;
the const char * re-spelling of ~2,500 format arguments.

Trees touched: symbolized/{app,net}, named, net/named, readable_sources/app/{g1,library},
app/src, verified/src, plus new recon/headers/g1_log.h. Nothing in tools/, board/,
emulator/, generated/, symbols/, any build file or linker script.

+++ recon/symbolized/app/debug_print.c
@@ -14,6 +14,7 @@
 
 #include <stdarg.h>
 #include <stdint.h>
+#include "../../headers/g1_log.h"
 
 extern uintptr_t ancs_get_conn_ctx(void);
 extern void memset_bytes(void *dst, int val, unsigned int len);
@@ -21,7 +22,7 @@ extern void vsnprintf_impl(void *dst, unsigned int size, unsigned int fmt, va_li
 extern void enqueue_debug(void *buf);
 extern void process_sync_buffer(uintptr_t arg);
 
-void debug_print(unsigned int format, ...)
+void debug_print(uintptr_t format, ...)
 {
     va_list arguments;
     va_start(arguments, format);
